Add files via upload
Added 'recover-on-crash' feature, that let merge all ts files on crash

+++ save-stream.py
@@ -3,7 +3,25 @@ import m3u8
 import streamlink
 import subprocess
 import os
-import datetime;
+import datetime
+
+def merge_all(tmp_folder,output_file):
+    txt_list = []
+    for i in os.listdir(tmp_folder):
+        txt_list.append('file ' + i + '\n')
+
+    list = open(tmp_folder + '/concat.txt', 'w+')
+    list.writelines(txt_list)
+    list.close()
+
+    print('[INFO] Stream downloaded. Merging all files to ' + output_file)
+    os.chdir(tmp_folder)
+    subprocess.call('ffmpeg -f concat -safe 0 -i concat.txt -c copy ../' + output_file + ' -hide_banner -loglevel panic')
+    print('[INFO] Merge successful. Removing tmp files')
+    os.chdir('../')
+    for i in os.listdir(tmp_folder):
+        os.remove(tmp_folder + '/' + i)
+    print('[INFO] Successfully deleted')
 
 def record_stream(url,iterations,tmp_folder,output_file):
     start_time = ts = datetime.datetime.now().timestamp()
@@ -18,29 +36,13 @@ def record_stream(url,iterations,tmp_folder,output_file):
         filedata = urllib.request.urlopen(str_uri).read()
 
         if filedata != prev_data:
-            f = open(tmp_folder+'/tmp-'+str(counter)+'.ts', 'ab+')
+            f = open(tmp_folder+'/tmp-'+str(counter)+'.ts', 'wb+')
             f.write(filedata)
             f.close()
             prev_data = filedata
             counter+=1
             print('[INFO] Writing segment '+str(counter)+'. File path - '+tmp_folder+'/tmp-'+str(counter)+'.ts. Left '+str(iterations-counter)+' segments or '+str((iterations-counter)*5)+' secs')
-
-    txt_list = []
-    for i in os.listdir(tmp_folder):
-        txt_list.append('file '+i+'\n')
-
-    list = open(tmp_folder+'/concat.txt','w+')
-    list.writelines(txt_list)
-    list.close()
-
-    print('[INFO] Stream downloaded. Merging all files to '+output_file)
-    os.chdir(tmp_folder)
-    subprocess.call('ffmpeg -f concat -safe 0 -i concat.txt -c copy ../'+output_file+' -hide_banner -loglevel panic')
-    print('[INFO] Merge successful. Removing tmp files')
-    os.chdir('../')
-    for i in os.listdir(tmp_folder):
-        os.remove(tmp_folder+'/'+i)
-    print('[INFO] Successfully deleted')
+    merge_all(tmp_folder,output_file)
     print('[INFO] Recording time - '+str(datetime.datetime.now().timestamp() - start_time)+'s ')
 
 print("[+] Enter url: ")
@@ -62,12 +64,17 @@ try:
 except:
     print('[INFO] Not creating folder ./result - it already exists')
 if os.listdir('stream-tmp'):
-    print('[INFO] Something found in ./stream-tmp folder. Deleting...')
-    for i in os.listdir('stream-tmp'):
-        os.remove('./stream-tmp/'+i)
-    print('[INFO] Successfully deleted')
+    print('[INFO] Something found in ./stream-tmp folder. Try to merge files into mp4, or Remove?[Y/n]')
+    input = input()
+    if input.capitalize() == 'Y' or input.capitalize() == 'YES' or input == '':
+        merge_all('stream-tmp', 'result/recovered-stream.mp4')
+    else:
+        for i in os.listdir('stream-tmp'):
+            os.remove('./stream-tmp/'+i)
+        print('[INFO] Successfully deleted')
 
 print('[INFO] Starting stream recorder')
 
 ts = datetime.datetime.now().timestamp()
+
 record_stream(uri,iter,'stream-tmp','result/stream-'+str(round(ts))+'.mp4')
